AGENT-975: Add support for agent-based installer minimal ISO
Configure the installer to create a minimal ISO. Add support
for a nimimal ISO installation in a disconnected environment
using the bootArtifactsBaseURL field. Made the PXE server into
a more generic boot server that can be used to server the rootfs
for minimal ISO.

+++ agent/cleanup.sh
@@ -15,17 +15,16 @@ early_cleanup_validation
 rm -rf "${OCP_DIR}/manifests"
 rm -rf "${OCP_DIR}/output"
 
-case "${AGENT_E2E_TEST_BOOT_MODE}" in
-  "PXE" )
-    sudo pkill agentpxeserver || true
-    rm -rf ${WORKING_DIR}/boot-artifacts
-    ;;
-  "DISKIMAGE" )
+if [[ "${AGENT_E2E_TEST_BOOT_MODE}" == "DISKIMAGE" ]]; then
     sudo rm -rf "${OCP_DIR}/cache"
     sudo rm -rf "${OCP_DIR}/temp"
     sudo podman rmi -f ${APPLIANCE_IMAGE} || true
-    ;;
-esac
+fi
+
+if [[ -d ${BOOT_SERVER_DIR} ]]; then
+   sudo pkill agentpxeserver || true
+   rm -rf ${BOOT_SERVER_DIR}
+fi
 
 sudo podman rm -f extlb || true
 sudo rm ${WORKING_DIR}/haproxy.* || true
